Rename exceptions to azurectl_exceptions
Avoid confusion about generic python exception class
