+++ azure_cli/container.py
@@ -14,7 +14,7 @@
 from azure.storage import BlobService
 
 # project
-from exceptions import *
+from azurectl_exceptions import *
 from logger import Logger
