+++ azure_cli/disk_task.py
@@ -37,7 +37,7 @@ from cli_task import CliTask
 from azure_account import AzureAccount
 from data_collector import DataCollector
 from logger import Logger
-from exceptions import *
+from azurectl_exceptions import *
 from disk import Disk
 from container import Container
