+++ azure_cli/image.py
@@ -15,7 +15,7 @@ from tempfile import NamedTemporaryFile
 from azure.servicemanagement import ServiceManagementService
 
 # project
-from exceptions import *
+from azurectl_exceptions import *
 from logger import Logger
