+++ test/unit/image_test.py
@@ -3,7 +3,7 @@ import mock
 from mock import patch
 from nose.tools import *
 from azure_cli.azure_account import AzureAccount
-from azure_cli.exceptions import *
+from azure_cli.azurectl_exceptions import *
 from azure_cli.image import Image
 
 import azure_cli
